refactor: extract shared match guards and team-name helper
